# packages/server/openapi.json
{"openapi":"3.0.2","paths":{"/api/total-article":{"get":{"summary":"Get total number of articles","tags":["chart"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/article-line-chart":{"get":{"tags":["chart"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/total-authors":{"get":{"tags":["chart"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/total-affiliations":{"get":{"tags":["chart"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/pie-chart":{"get":{"tags":["chart"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/articles":{"get":{"tags":["article"],"parameters":[],"responses":{"200":{"description":"200"}}}},"/api/crawl":{"post":{"tags":["article"],"parameters":[],"requestBody":{"description":"Body","content":{"application/json":{"schema":{"type":"object","properties":{"url":{"type":"string","format":"uri"},"authorId":{"type":"string","minLength":1}},"required":["url","authorId"]}}}},"responses":{"201":{"description":"201"}}}},"/api/authors":{"get":{"tags":["author"],"parameters":[],"responses":{"200":{"description":"200"}}},"post":{"tags":["author"],"parameters":[],"requestBody":{"description":"Body","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"url":{"type":"string","format":"uri"},"affiliation":{"nullable":true}},"required":["name","url"]}}}},"responses":{"201":{"description":"201"},"404":{"description":"404"}}}},"/api/authors/{id}":{"patch":{"tags":["author"],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"description":"Body","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string"},"url":{"type":"string","format":"uri"},"schedule":{"type":"number"},"articles":{"type":"array","items":{"type":"string"}},"affiliation":{"type":"array","items":{"type":"string"}}}}}}},"responses":{"200":{"description":"200"}}},"delete":{"tags":["author"],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200"}}}},"/api/affiliation":{"post":{"tags":["affiliation"],"parameters":[],"requestBody":{"description":"Body","content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","minLength":2},"parent":{"type":"string"},"admin":{"type":"object","properties":{"email":{"type":"string","format":"email"},"name":{"type":"string","minLength":2},"password":{"type":"string","minLength":8}},"required":["email"]}},"required":["name"]}}}},"responses":{"201":{"description":"201","content":{"application/json":{"schema":{"nullable":true}}}}}}},"/api/affiliation/{id}":{"get":{"tags":["affiliation"],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"type":"object","properties":{"_id":{"nullable":true},"name":{"type":"string"},"parent":{"type":"string"},"authors":{"type":"array","items":{"nullable":true}},"users":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"emailVerified":{"type":"boolean"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["_id","name","email","emailVerified","createdAt","updatedAt"]}},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["name","authors","createdAt","updatedAt"]}}}}}}},"/api/raw-affiliations":{"get":{"tags":["affiliation"],"parameters":[],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"nullable":true},"name":{"type":"string"},"parent":{"nullable":true},"authors":{"type":"array","items":{"nullable":true}},"users":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"emailVerified":{"type":"boolean"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["_id","name","email","emailVerified","createdAt","updatedAt"]}},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["name","authors","createdAt","updatedAt"]}}}}}}}},"/api/affiliations":{"get":{"tags":["affiliation"],"parameters":[],"responses":{"200":{"description":"200","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"nullable":true},"name":{"type":"string"},"parent":{"type":"string"},"authors":{"type":"array","items":{"nullable":true}},"users":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"emailVerified":{"type":"boolean"},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["_id","name","email","emailVerified","createdAt","updatedAt"]}},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["name","authors","createdAt","updatedAt"]}}}}}}}}},"info":{"title":"Posts API","version":"1.0.0"}}